+++ Ds3/Ds3.csproj
@@ -578,9 +578,9 @@
     <Compile Include="Calls\VerifyTapeSpectraS3Response.cs" />
     <Compile Include="Calls\VerifyUserIsMemberOfGroupSpectraS3Request.cs" />
     <Compile Include="Calls\VerifyUserIsMemberOfGroupSpectraS3Response.cs" />
-    <Compile Include="Helpers\Ds3Diagnostics\CacheNearCapacity.cs" />
-    <Compile Include="Helpers\Ds3Diagnostics\NoPools.cs" />
-    <Compile Include="Helpers\Ds3Diagnostics\NoTapes.cs" />
+    <Compile Include="Helpers\Ds3Diagnostics\CacheNearCapacityDiagnostic.cs" />
+    <Compile Include="Helpers\Ds3Diagnostics\NoPoolsDiagnostic.cs" />
+    <Compile Include="Helpers\Ds3Diagnostics\NoTapesDiagnostic.cs" />
     <Compile Include="Helpers\Ds3Diagnostics\DiagnosticsMessages.Designer.cs">
       <DependentUpon>DiagnosticsMessages.resx</DependentUpon>
       <AutoGen>True</AutoGen>
@@ -590,8 +590,8 @@
     <Compile Include="Helpers\Ds3Diagnostics\Ds3DiagnosticResult.cs" />
     <Compile Include="Helpers\Ds3Diagnostics\Ds3Diagnostic.cs" />
     <Compile Include="Helpers\Ds3Diagnostics\IDs3DiagnosticCheck.cs" />
-    <Compile Include="Helpers\Ds3Diagnostics\OfflineTapes.cs" />
-    <Compile Include="Helpers\Ds3Diagnostics\PoweredOffPools.cs" />
+    <Compile Include="Helpers\Ds3Diagnostics\OfflineTapesDiagnostic.cs" />
+    <Compile Include="Helpers\Ds3Diagnostics\PoweredOffPoolsDiagnostic.cs" />
     <Compile Include="Helpers\Ds3JobOptions.cs" />
     <Compile Include="Helpers\Ds3ReadJobOptions.cs" />
     <Compile Include="Helpers\Ds3WriteJobOptions.cs" />
